+++ examples/multidimgraph_2d.jl
@@ -1,6 +1,5 @@
 using Graphs: grid, has_edge, has_vertex, nv
-using NamedGraphs.GraphsExtensions: subgraph
-using NamedGraphs: NamedGraph, ⊔
+using NamedGraphs: NamedGraph, subgraph, ⊔
 
 simple_graph = grid((2, 2))
 vs = [("X", 1), ("X", 2), ("Y", 1), ("Y", 2)]