+++ test/test_exports.jl
@@ -63,27 +63,17 @@ using Test: @test, @testset
     @testset "GraphsExtensions" begin
         exports = [
             :GraphsExtensions,
-            :add_edges,
-            :add_edges!,
-            :add_vertices,
             :boundary_edges,
             :convert_vertextype,
             :default_root_vertex,
-            :directed_graph,
-            :forest_cover,
-            :forest_cover_edge_sequence,
             :in_incident_edges,
             :incident_edges,
             :is_leaf_vertex,
             :leaf_vertices,
             :post_order_dfs_edges,
             :post_order_dfs_vertices,
-            :rem_edges,
-            :rem_edges!,
-            :rem_vertices,
             :similar_graph,
             :subgraph,
-            :undirected_graph,
             :vertextype,
         ]
         @test issetequal(names(NamedGraphs.GraphsExtensions), exports)