+++ test/test_multidimgraph.jl
@@ -1,6 +1,5 @@
 using Graphs: add_edge!, add_vertex!, grid, has_edge, has_vertex, ne, nv
-using NamedGraphs.GraphsExtensions: subgraph
-using NamedGraphs: NamedGraph, disjoint_union, ⊔
+using NamedGraphs: NamedGraph, disjoint_union, subgraph, ⊔
 using Test: @test, @testset
 
 @testset "NamedGraph" begin